int16_t/int32_t instead of short/long

+++ audio_read.c
@@ -24,7 +24,7 @@ unsigned long read_samples (FILE * musicin, long int *sample_buffer,
   unsigned long samples_read =0 ;
   static unsigned long samples_to_read=0;
   static char init = TRUE;
-  short pcm_sample_buffer[9216];	/*for correct reading of pcm-data */
+  int16_t pcm_sample_buffer[9216];	/*for correct reading of pcm-data */
   int i;
   
   //printf("AIFF = %i \n", *aiff); // hack
@@ -56,7 +56,7 @@ unsigned long read_samples (FILE * musicin, long int *sample_buffer,
   } else {
     
     if ((samples_read =
-	 fread (pcm_sample_buffer, sizeof (short), (int) samples_read,
+	 fread (pcm_sample_buffer, sizeof (int16_t), (int) samples_read,
 		musicin)) == 0)
       if (verbosity >= 2)
 	printf ("Hit end of audio data (pcm)\n");
@@ -83,7 +83,7 @@ unsigned long read_samples_new (FILE * musicin, long int *sample_buffer,
   unsigned long samples_read;
   static unsigned long samples_to_read;
   static char init = TRUE;
-  short pcm_sample_buffer[9216];	/*for correct reading of pcm-data */
+  int16_t pcm_sample_buffer[9216];	/*for correct reading of pcm-data */
   int i;
 
   if (init) {
@@ -97,7 +97,7 @@ unsigned long read_samples_new (FILE * musicin, long int *sample_buffer,
 
  
   if ((samples_read =
-       fread (pcm_sample_buffer, sizeof (short), (int) samples_read,
+       fread (pcm_sample_buffer, sizeof (int16_t), (int) samples_read,
 	      musicin)) == 0)
     if (verbosity >= 2)
       printf ("Hit end of audio data\n");
@@ -153,7 +153,7 @@ get_audio (FILE * musicin,
   )
 {
   int k, j, i;
-  long insamp[9216];
+  int32_t insamp[9216];
   unsigned long samples_read;
   int lay;
   int lfe;
@@ -270,7 +270,7 @@ get_audio_new (FILE * musicin,
   )
 {
   int j, i;
-  long insamp[9216];
+  int32_t insamp[9216];
   unsigned long samples_read=0;
   //MFC  int lay;
   int lfe;
@@ -296,7 +296,7 @@ get_audio_new (FILE * musicin,
 #else
   if (stereo == 2) {
     samples_read = read_samples (musicin, insamp, num_samples,
-				 (unsigned long) ((2 + lfe) * 1152),
+				 (uint32_t) ((2 + lfe) * 1152),
 				 byte_per_sample, aiff);
     for (j = 0; j < 1152; j++) {	/* fixed bug 28.6.93 S.R. */
       buffer[0][j] = insamp[(2 + lfe) * j];
@@ -309,7 +309,7 @@ get_audio_new (FILE * musicin,
     }
   } else {			/* layer 2 (or 3), mono */
     samples_read = read_samples (musicin, insamp, num_samples,
-				 (unsigned long) 1152, byte_per_sample,
+				 (uint32_t) 1152, byte_per_sample,
 				 aiff);
     for (j = 0; j < 1152; j++) {
       buffer[0][j] = insamp[j];
